tests: move ChangeIdIndex test from test_revset to test_index
